async
For this lab,we search through a set of files for a given string.
Create a directory, and place a number of files inside, with some text
in each. In some of the files, place the word "needle." Then, using the
async module and streams, look through the files for "needle,"
hopefully reading only as much of the file as necessary (i.e., no more
than needed to find the search term).